added ros params
