+++ src/world_modeling/occupancy_segmentation/package.xml
@@ -2,9 +2,9 @@
 <package format="3">
   <name>occupancy_segmentation</name>
   <version>0.0.0</version>
-  <description>A sample ROS package for pubsub communication</description>
+  <description>A ROS2 package to handle LiDAR ground segmentation using Patchworks</description>
 
-  <maintainer email="cmahalingam@watonomous.ca">Cheran Mahalingam</maintainer>
+  <maintainer email="akilpath@watonomous.ca">Akil Pathiranage</maintainer>
   <license>TODO</license>
 
   <!--https://www.ros.org/reps/rep-0149.html#dependency-tags-->
